refactor: move exec_promise and util to util/
