+++ src/types.ts
@@ -1,4 +1,4 @@
-import { Attachment, DocumentID } from '@dittolive/ditto'
+import { Attachment, Document, DocumentID } from '@dittolive/ditto'
 import {
     DEFAULT_DESCRIPTION,
     DEFAULT_NODE_ID,
@@ -27,4 +27,17 @@ export const defaultPayload: Payload = {
     state: DEFAULT_STATE,
     isRemoved: false,
     siteId: '',
+}
+
+// TODO probably not needed
+export function deserPayload(doc: Document): Record<string, string> {
+    const payload: Payload = defaultPayload
+    const fields: Record<string, string> = {}
+    Object.keys(payload).map((key) => {
+        const val = doc.at(key)
+        if (val != undefined && val != null) {
+            fields[key] = val.value.toString()
+        }
+    })
+    return fields
 }
